@@ -1,6 +1,6 @@
 module github.com/evrone/go-clean-template
 
-go 1.25
+go 1.26
 
 tool (
 	github.com/daixiang0/gci
